Migrate legacy ordered waypoints parsing
