+++ src/main/java/de/hysky/skyblocker/skyblock/waypoint/OrderedWaypoints.java
@@ -50,25 +50,12 @@ import static net.fabricmc.fabric.api.client.command.v2.ClientCommandManager.lit
 @Deprecated
 public class OrderedWaypoints {
 	private static final Logger LOGGER = LogUtils.getLogger();
-	@VisibleForTesting
 	static final Codec<Map<String, OrderedWaypointGroup>> SERIALIZATION_CODEC = Codec.unboundedMap(Codec.STRING, OrderedWaypointGroup.CODEC).xmap(Object2ObjectOpenHashMap::new, Object2ObjectOpenHashMap::new);
-	@VisibleForTesting
-	static final String PREFIX = "[Skyblocker::OrderedWaypoints::v1]";
 	public static final Path PATH = SkyblockerMod.CONFIG_DIR.resolve("ordered_waypoints.json");
 
 	@Init
 	public static void init() {
 		ClientLifecycleEvents.CLIENT_STARTED.register(_client -> load());
-		ClientCommandRegistrationCallback.EVENT.register(OrderedWaypoints::registerCommands);
-	}
-
-	private static void registerCommands(CommandDispatcher<FabricClientCommandSource> dispatcher, CommandRegistryAccess registryAccess) {
-		dispatcher.register(literal(SkyblockerMod.NAMESPACE)
-				.then(literal("waypoints")
-						.then(literal("ordered")
-								.then(literal("import")
-										.then(literal("skyblocker")
-												.executes(context -> fromSkyblockerFormat(context.getSource())))))));
 	}
 
 	/**
@@ -90,41 +77,12 @@ public class OrderedWaypoints {
 	}
 
 	/**
-	 * @deprecated Use {@link Waypoints} instead.
-	 */
-	@Deprecated
-	private static int fromSkyblockerFormat(FabricClientCommandSource source) {
-		try {
-			String importCode = MinecraftClient.getInstance().keyboard.getClipboard();
-
-			if (importCode.startsWith(PREFIX)) {
-				String encoded = importCode.replace(PREFIX, "");
-				byte[] decoded = Base64.getDecoder().decode(encoded);
-
-				String json = new String(new GZIPInputStream(new ByteArrayInputStream(decoded)).readAllBytes());
-				Map<String, OrderedWaypointGroup> importedWaypoints = SERIALIZATION_CODEC.parse(JsonOps.INSTANCE, JsonParser.parseString(json)).getOrThrow();
-
-				migrateOrderedWaypoints(importedWaypoints);
-				source.sendFeedback(Constants.PREFIX.get().append(Text.translatable("skyblocker.waypoints.ordered.import.skyblocker.success")));
-			} else {
-				source.sendError(Constants.PREFIX.get().append(Text.translatable("skyblocker.waypoints.ordered.import.skyblocker.unknownFormatHeader")));
-			}
-		} catch (Exception e) {
-			LOGGER.error("[Skyblocker Ordered Waypoints] Failed to import waypoints!", e);
-			source.sendError(Constants.PREFIX.get().append(Text.translatable("skyblocker.waypoints.ordered.import.skyblocker.fail")));
-		}
-
-		return Command.SINGLE_SUCCESS;
-	}
-
-	/**
 	 * Migrates the given ordered waypoints to waypoints.
 	 */
 	private static void migrateOrderedWaypoints(Map<String, OrderedWaypointGroup> orderedWaypoints) {
 		toWaypointGroups(orderedWaypoints.values()).forEach(Waypoints::putWaypointGroup);
 	}
 
-	@VisibleForTesting
 	static List<WaypointGroup> toWaypointGroups(Collection<OrderedWaypointGroup> orderedWaypoints) {
 		return orderedWaypoints.stream().flatMap(orderedWaypointGroup -> Stream.of(
 				new WaypointGroup(orderedWaypointGroup.name, Location.DWARVEN_MINES, orderedWaypointGroup.waypoints.stream().map(waypoint -> new OrderedNamedWaypoint(waypoint.pos, "", new float[]{0, 1, 0})).collect(Collectors.toList()), true),
@@ -133,7 +91,6 @@ public class OrderedWaypoints {
 	}
 
 	@Deprecated
-	@VisibleForTesting
 	record OrderedWaypointGroup(String name, boolean enabled, ObjectArrayList<OrderedWaypoint> waypoints) {
 		static final Codec<OrderedWaypointGroup> CODEC = RecordCodecBuilder.create(instance -> instance.group(
 				Codec.STRING.fieldOf("name").forGetter(OrderedWaypointGroup::name),
